fix(openclaw): isolate config restore ownership
