+++ src/lib/state/sandbox.ts
@@ -31,6 +31,7 @@ import { OPENSHELL_PROBE_TIMEOUT_MS } from "../adapters/openshell/timeouts.js";
 import type { AgentStateFile } from "../agent/defs.js";
 import { loadAgent } from "../agent/defs.js";
 import { isRecord, type UnknownRecord } from "../core/json-types.js";
+import { mergeOpenClawRestoredConfig } from "./openclaw-config-merge.js";
 import { shellQuote } from "../runner.js";
 import { isSensitiveFile, sanitizeConfigFile } from "../security/credential-filter.js";
 import * as registry from "./registry.js";
@@ -920,135 +921,6 @@ function readCurrentStateFile(
     _log(`WARNING: state file current read ${spec.path} failed: ${detail.substring(0, 200)}`);
   }
   return null;
-}
-
-function isPlainJsonObject(value: unknown): value is Record<string, unknown> {
-  return isRecord(value);
-}
-
-function cloneJson<T>(value: T): T {
-  if (value === undefined) return undefined as T;
-  return JSON.parse(JSON.stringify(value)) as T;
-}
-
-function mergeJsonObjects(
-  base: Record<string, unknown>,
-  overlay: Record<string, unknown>,
-): Record<string, unknown> {
-  const merged: Record<string, unknown> = cloneJson(base);
-  for (const [key, value] of Object.entries(overlay)) {
-    const existing = merged[key];
-    if (isPlainJsonObject(existing) && isPlainJsonObject(value)) {
-      merged[key] = mergeJsonObjects(existing, value);
-    } else {
-      merged[key] = cloneJson(value);
-    }
-  }
-  return merged;
-}
-
-const NEMOCLAW_MANAGED_OPENCLAW_CHANNELS = new Set([
-  "discord",
-  "slack",
-  "telegram",
-  "whatsapp",
-  "wechat",
-  "openclaw-weixin",
-]);
-
-function mergeOpenClawChannels(
-  backupChannels: unknown,
-  currentChannels: unknown,
-): Record<string, unknown> | unknown {
-  if (!isPlainJsonObject(backupChannels)) return cloneJson(currentChannels);
-  if (!isPlainJsonObject(currentChannels)) return cloneJson(backupChannels);
-
-  const merged: Record<string, unknown> = cloneJson(currentChannels);
-  for (const [key, value] of Object.entries(backupChannels)) {
-    if (key === "defaults") {
-      merged[key] =
-        isPlainJsonObject(value) && isPlainJsonObject(merged[key])
-          ? mergeJsonObjects(merged[key] as Record<string, unknown>, value)
-          : cloneJson(value);
-      continue;
-    }
-
-    if (NEMOCLAW_MANAGED_OPENCLAW_CHANNELS.has(key)) {
-      // Freshly generated channel blocks carry current OpenShell placeholder
-      // revisions and current start/stop/add/remove state. Never resurrect a
-      // managed channel that the fresh config omitted, and never overwrite a
-      // present managed channel with a stale backed-up account block.
-      continue;
-    }
-
-    const existing = merged[key];
-    merged[key] =
-      isPlainJsonObject(existing) && isPlainJsonObject(value)
-        ? mergeJsonObjects(existing, value)
-        : cloneJson(value);
-  }
-  return merged;
-}
-
-function mergeOpenClawModels(backupModels: unknown, currentModels: unknown): unknown {
-  if (!isPlainJsonObject(backupModels)) return cloneJson(currentModels);
-  if (!isPlainJsonObject(currentModels)) return cloneJson(backupModels);
-
-  const merged = mergeJsonObjects(currentModels, backupModels);
-  const backupProviders = backupModels.providers;
-  const currentProviders = currentModels.providers;
-  if (isPlainJsonObject(backupProviders) && isPlainJsonObject(currentProviders)) {
-    merged.providers = {
-      ...cloneJson(backupProviders),
-      // Current generated provider entries win so rebuild does not restore stale
-      // runtime placeholders or model routing for providers NemoClaw manages.
-      ...cloneJson(currentProviders),
-    };
-  }
-  return merged;
-}
-
-function mergeOpenClawPlugins(backupPlugins: unknown, currentPlugins: unknown): unknown {
-  if (!isPlainJsonObject(backupPlugins)) return cloneJson(currentPlugins);
-  if (!isPlainJsonObject(currentPlugins)) return cloneJson(backupPlugins);
-
-  const merged = mergeJsonObjects(currentPlugins, backupPlugins);
-  const backupEntries = backupPlugins.entries;
-  const currentEntries = currentPlugins.entries;
-  if (isPlainJsonObject(backupEntries) && isPlainJsonObject(currentEntries)) {
-    merged.entries = {
-      ...cloneJson(backupEntries),
-      // Current generated plugin enablement wins for channels/provider plugins;
-      // backup-only custom plugin entries are still preserved.
-      ...cloneJson(currentEntries),
-    };
-  }
-  return merged;
-}
-
-export function mergeOpenClawRestoredConfig(
-  backedUpConfig: unknown,
-  currentConfig: unknown,
-): unknown {
-  if (!isPlainJsonObject(backedUpConfig)) return cloneJson(currentConfig ?? backedUpConfig);
-  if (!isPlainJsonObject(currentConfig)) return cloneJson(backedUpConfig);
-
-  const merged = mergeJsonObjects(currentConfig, backedUpConfig);
-
-  // Runtime-owned sections are regenerated during rebuild and must survive the
-  // restore. The backup is sanitized (gateway removed) and may contain stale
-  // OpenShell resolve placeholder revisions, so it is unsafe to wholesale
-  // replace the fresh config with the backed-up file.
-  for (const key of ["gateway", "proxy", "diagnostics"] as const) {
-    if (key in currentConfig) merged[key] = cloneJson(currentConfig[key]);
-    else delete merged[key];
-  }
-
-  merged.channels = mergeOpenClawChannels(backedUpConfig.channels, currentConfig.channels);
-  merged.models = mergeOpenClawModels(backedUpConfig.models, currentConfig.models);
-  merged.plugins = mergeOpenClawPlugins(backedUpConfig.plugins, currentConfig.plugins);
-
-  return merged;
 }
 
 function shouldMergeOpenClawConfig(
